API with two routes created
One route GETs all of the prompt templates available in the server. The other one is meant for upload. It's already working, it takes .mp3s stores them in a tmp folder, and logs their path in a sqlite database. The conversion from .mp4s to .mp3s should be made in the frontend.